Add check python and git

+++ toolkit.sh
@@ -9,23 +9,33 @@ fi
 echo "Usage: $0 with token $1"
 GITHUB_TOKEN=$1
 
-# Install Python
+# # Check if Python is installed
 PYTHON_VERSION="3.12.6"
 PYTHON_BIN="/usr/local/bin/python${PYTHON_VERSION%.*}"
-echo "Installing Python..."
-wget "https://www.python.org/ftp/python/${PYTHON_VERSION}/Python-${PYTHON_VERSION}.tgz"
-tar xvf Python-${PYTHON_VERSION}.tgz
-cd Python-${PYTHON_VERSION}/
-./configure --enable-optimizations
-make altinstall
-${PYTHON_BIN} -m ensurepip --default-pip
-${PYTHON_BIN} -m pip install --upgrade pip setuptools wheel
-cd
-rm -rf Python-${PYTHON_VERSION}*
 
-# Install Git
-echo "Installing Git..."
-apt-get install -y git
+if python3 --version &>/dev/null; then
+    echo "Python is already installed."
+else
+    echo "Python not found. Installing Python..."
+    wget "https://www.python.org/ftp/python/${PYTHON_VERSION}/Python-${PYTHON_VERSION}.tgz"
+    tar xvf Python-${PYTHON_VERSION}.tgz
+    cd Python-${PYTHON_VERSION}/
+    ./configure --enable-optimizations
+    make altinstall
+    ${PYTHON_BIN} -m ensurepip --default-pip
+    ${PYTHON_BIN} -m pip install --upgrade pip setuptools wheel
+    cd ..
+    rm -rf Python-${PYTHON_VERSION}*
+fi
+
+# Check if Git is installed
+if git --version &>/dev/null; then
+    echo "Git is already installed."
+else
+    echo "Git not found. Installing Git..."
+    apt-get update
+    apt-get install -y git
+fi
 
 # Clone toolkit-installer repository using the provided token
 echo "Cloning toolkit-installer repository..."
